Move navigation component to `app` module
Moving `ScreenHostView` & `ScreenNavigator`
