+++ router/build.gradle
@@ -31,10 +31,6 @@ dependencies {
   implementation "io.reactivex.rxjava2:rxjava:$versions.rxJava"
   implementation "com.jakewharton.timber:timber:$versions.timber"
 
-  // If we add navigation-fragment-ktx or navigation-ui-ktx in future, we can remove this.
-  // Because this gets added as a transitive dependency.
-  api "androidx.navigation:navigation-runtime-ktx:$versions.navigation"
-
   testImplementation "pl.pragmatists:JUnitParams:$versions.junitParams"
   testImplementation "com.google.truth:truth:$versions.truth"
 }
